feat: Sync entity movement with time

+++ src/network/datatypes/Data.java
@@ -34,15 +34,15 @@ abstract public class Data {
         if (this instanceof MovementData) {
             MovementData data = (MovementData) this;
             return String.join(",", DataType.movement.toString(), Integer.toString(data.getUserId()),
-                    data.getDirection().toString(), Integer.toString(data.getPlayerX()),
-                    Integer.toString(data.getPlayerY()));
+                    data.getDirection().toString(), Double.toString(data.getPlayerX()),
+                    Double.toString(data.getPlayerY()));
         }
 
         if (this instanceof ToggleFireData) {
             ToggleFireData data = (ToggleFireData) this;
             return String.join(",", DataType.fire.toString(), Integer.toString(data.getUserId()),
-                    data.getStatus().toString(), Integer.toString(data.getPlayerX()),
-                    Integer.toString(data.getPlayerY()));
+                    data.getStatus().toString(), Double.toString(data.getPlayerX()),
+                    Double.toString(data.getPlayerY()));
         }
 
         System.out.println("Error serializing data " + this.getClass().getSimpleName());
@@ -74,12 +74,12 @@ abstract public class Data {
 
         if (dataChunks[0].equals(DataType.movement.toString())) {
             return new MovementData(Integer.parseInt(dataChunks[1]), MovementData.Movement.valueOf(dataChunks[2]),
-                    Integer.parseInt(dataChunks[3]), Integer.parseInt(dataChunks[4]));
+                    Double.parseDouble(dataChunks[3]), Double.parseDouble(dataChunks[4]));
         }
 
         if (dataChunks[0].equals(DataType.fire.toString())) {
             return new ToggleFireData(Integer.parseInt(dataChunks[1]), ToggleFireData.Status.valueOf(dataChunks[2]),
-                    Integer.parseInt(dataChunks[3]), Integer.parseInt(dataChunks[4]));
+                    Double.parseDouble(dataChunks[3]), Double.parseDouble(dataChunks[4]));
         }
 
         System.out.println("Error deserializing data " + serializedData);
